Refactor: setup new naming scheme for tool call
