+++ app/tools/exercise/executor.py
@@ -1,19 +1,54 @@
 import logging
 from typing import List, Literal
 
-from app.tools.generate_exercise.vector_database import vector_client
+from app.tools.exercise.vector_database import vector_client
 from app.utils.llm_utils import async_groq_request, async_openai_request
-from app.tools.generate_exercise.models import RetrievedDocSchema
-from app.tools.generate_exercise.prompts import (
+from app.tools.exercise.models import RetrievedDocSchema
+from app.tools.exercise.prompts import (
     PIPELINE_QUESTION_GENERATOR_PROMPT,
     PIPELINE_QUESTION_GENERATOR_USER_PROMPT,
 )
-from app.tools.generate_exercise.modules import (
+from app.tools.exercise.modules import (
     chromadb_retriever,
     query_rewriter,
 )
 
 logger = logging.getLogger(__name__)
+
+
+async def generate_exercise(user_query: str):
+    # Rewrite the user query
+    original_query = user_query
+    rewritten_query = await query_rewriter(original_query, llm="llama3-8b-8192")
+
+    verbose = False
+
+    # Retrieve the relevant content and exercises
+    retrieved_content: List[RetrievedDocSchema] = await chromadb_retriever(
+        model_class=vector_client,
+        retrieval_msg=rewritten_query,
+        size=5,
+        doc_type="Content",
+        verbose=verbose,
+    )
+    retrieved_exercises: List[RetrievedDocSchema] = await chromadb_retriever(
+        model_class=vector_client,
+        retrieval_msg=rewritten_query,
+        size=2,
+        doc_type="Exercise",
+        verbose=verbose,
+    )
+
+    # Format the context and prompt
+    context = _format_context(retrieved_content, retrieved_exercises)
+    system_prompt = PIPELINE_QUESTION_GENERATOR_PROMPT.format()
+    user_prompt = PIPELINE_QUESTION_GENERATOR_USER_PROMPT.format(
+        query=user_query, context_str=context
+    )
+
+    # Generate a question based on the context
+    res = await _generate(system_prompt, user_prompt, model="llama3-70b-8192")
+    return res
 
 
 async def _generate(
@@ -102,38 +137,3 @@ def _format_context(
         context_parts.append(f"{doc.source.chunk}")
 
     return "\n".join(context_parts)
-
-
-async def exercise_generator(user_query: str):
-    # Rewrite the user query
-    original_query = user_query
-    rewritten_query = await query_rewriter(original_query, llm="llama3-8b-8192")
-
-    verbose = False
-
-    # Retrieve the relevant content and exercises
-    retrieved_content: List[RetrievedDocSchema] = await chromadb_retriever(
-        model_class=vector_client,
-        retrieval_msg=rewritten_query,
-        size=5,
-        doc_type="Content",
-        verbose=verbose,
-    )
-    retrieved_exercises: List[RetrievedDocSchema] = await chromadb_retriever(
-        model_class=vector_client,
-        retrieval_msg=rewritten_query,
-        size=2,
-        doc_type="Exercise",
-        verbose=verbose,
-    )
-
-    # Format the context and prompt
-    context = _format_context(retrieved_content, retrieved_exercises)
-    system_prompt = PIPELINE_QUESTION_GENERATOR_PROMPT.format()
-    user_prompt = PIPELINE_QUESTION_GENERATOR_USER_PROMPT.format(
-        query=user_query, context_str=context
-    )
-
-    # Generate a question based on the context
-    res = await _generate(system_prompt, user_prompt, model="llama3-70b-8192")
-    return res
